+++ RoomfulPublisherProject/Packages/net.roomful.assets/API/Serialisation/Components/TeleportWithMap.cs
@@ -1,0 +1,14 @@
+﻿using net.roomful.api.props;
+using UnityEngine;
+
+namespace net.roomful.assets.serialization
+{
+    public class TeleportWithMap : MonoBehaviour, IPropComponent, IRecreatableOnLoad
+    {
+        public void Init(IProp prop, int componentIndex) { }
+        public void OnPropUpdated() { }
+        public void PropScaleChanged() { }
+        public void OnZoomViewOpen() { }
+        public void OnZoomViewClosed() { }
+    }
+}